% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Best_Arthritis}
\alias{Best_Arthritis}
\title{Arthritis}
\format{An object of class \code{glm} (inherits from \code{lm}) of length 30.}
\usage{
Best_Arthritis
}
\description{
Arthritis
}
\keyword{datasets}
