% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Best_Binge_Drinker}
\alias{Best_Binge_Drinker}
\title{Binge Drinking}
\format{An object of class \code{glm} (inherits from \code{lm}) of length 30.}
\usage{
Best_Binge_Drinker
}
\description{
Binge Drinking
}
\keyword{datasets}
