% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Best_Depression}
\alias{Best_Depression}
\title{Depression}
\format{An object of class \code{glm} (inherits from \code{lm}) of length 30.}
\usage{
Best_Depression
}
\description{
Depression
}
\keyword{datasets}
