% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Best_High_Blood_Pressure}
\alias{Best_High_Blood_Pressure}
\title{Best High Blood Pressure}
\format{An object of class \code{glm} (inherits from \code{lm}) of length 30.}
\usage{
Best_High_Blood_Pressure
}
\description{
Best High Blood Pressure
}
\keyword{datasets}
