% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Best_High_Cholesterol}
\alias{Best_High_Cholesterol}
\title{High Cholestorol}
\format{An object of class \code{glm} (inherits from \code{lm}) of length 30.}
\usage{
Best_High_Cholesterol
}
\description{
High Cholestorol
}
\keyword{datasets}
