% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Black_Older}
\alias{Black_Older}
\title{Black Older Interaction}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49 rows and 2 columns.}
\usage{
Black_Older
}
\description{
Black Older Interaction
}
\keyword{datasets}
