% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Census_data}
\alias{Census_data}
\title{Import census data to save time}
\format{An object of class \code{data.frame} with 49 rows and 23 columns.}
\usage{
Census_data
}
\description{
Import census data to save time
}
\keyword{datasets}
