% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{df_final}
\alias{df_final}
\title{Final Data Frame}
\format{An object of class \code{data.frame} with 10189 rows and 30 columns.}
\usage{
df_final
}
\description{
Final Data Frame
}
\keyword{datasets}
