% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Education_tract}
\alias{Education_tract}
\title{Education Tract}
\format{An object of class \code{data.frame} with 294 rows and 5 columns.}
\usage{
Education_tract
}
\description{
Education Tract
}
\keyword{datasets}
