% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{eville}
\alias{eville}
\title{#' Census Key
#' @importFrom tidycensus census_api_key get_acs
census_api_key("94bc4d9ee635889131819300fb54c51a0a92b05f", install = FALSE)}
\format{An object of class \code{data.frame} with 49 rows and 57 columns.}
\usage{
eville

eville

eville
}
\description{
#' Education Tract
Education_tract <- data.frame(get_acs(geography = "tract",
                                      geometry = FALSE,
                                      state = 18,
                                      county = "163",
                                      variables = c(education_num_counted = "B06009_001",
                                                    No_GED = "B06009_002",
                                                    High_School_Grad = "B06009_003",
                                                    Some_college_or_associates = "B06009_004",
                                                    Bachelors_degree = "B06009_005",
                                                    Professional_degree = "B06009_006"
                                                    )
                                      )
                              )
#' Sorting Education Data
No_GED <- Education_tract %>%
  filter(variable == "No_GED") %>%
  select(GEOID, no_GED = estimate)
}
\details{
#' High School
High_School <- Education_tract %>%
  filter(variable == "High_School_Grad") %>%
  select(GEOID, high_school_grad = estimate)

#' College
College <- Education_tract %>%
  filter(variable %in% c("Some_college_or_associates", "Bachelors_degree", "Professional_degree")) %>%
  group_by(GEOID) %>%
  summarise(college = sum(estimate))

#' Full Census
Census_data <- No_GED %>%
  left_join(High_School, by = "GEOID") %>%
  left_join(College, by = "GEOID")

#' Sex Data
Sex_tract <- data.frame(get_acs(geography = "tract",
                                geometry = FALSE,
                                state = 18,
                                county = "163",
                                variables = c(Males = "B01001_002")
                                )
                        ) %>% select(GEOID, males = estimate)
#' Joining Sex
Census_data <- Census_data %>% left_join(Sex_tract, by = "GEOID")

#' Income Data
Income_tract <- data.frame(get_acs(geography = "tract",
                                   geometry = FALSE,
                                   state = 18,
                                   county = "163",
                                   variables = c(income_00_10 = "B19001_002",
                                                 income_10_15 = "B19001_003",
                                                 income_15_20 = "B19001_004",
                                                 income_20_25 = "B19001_005",
                                                 income_25_30 = "B19001_006",
                                                 income_30_35 = "B19001_007",
                                                 income_35_40 = "B19001_008",
                                                 income_40_45 = "B19001_009",
                                                 income_45_50 = "B19001_010",
                                                 income_50_60 = "B19001_011",
                                                 income_60_75 = "B19001_012",
                                                 income_75_100 = "B19001_013",
                                                 income_100_125 = "B19001_014",
                                                 income_125_150 = "B19001_015",
                                                 income_150_200 = "B19001_016",
                                                 income_200_up = "B19001_017"
                                                 )
                                   )
                           )
#' Poverty
Poverty <- Income_tract %>%
  filter(variable %in% c("income_00_10", "income_10_15", "income_15_20", "income_20_25")) %>%
  group_by(GEOID) %>%
  summarise(poverty = sum(estimate))

#' Middle Class
Middle_Class <- Income_tract %>% filter(variable %in% c("income_25_30","income_30_35","income_35_40","income_40_45","income_45_50","income_50_55","income_55_60","income_60_65","income_65_70","income_70_75")) %>%
  group_by(GEOID) %>%
  summarise(middle_class = sum(estimate))

#' Upper Class
Upper_Class <- Income_tract %>% filter(variable %in% c("income_75_100","income_100_125","income_125_150","income_150_200","income_200_up")) %>%
  group_by(GEOID) %>%
  summarise(upper_class = sum(estimate))

#' Census Join Income
Census_data <- Census_data %>%
  left_join(Poverty, by = "GEOID") %>%
  left_join(Middle_Class, by = "GEOID") %>%
  left_join(Upper_Class, by = "GEOID")

#' Age Data
Age_data <- data.frame(get_acs(geography = "tract",
                               geometry = FALSE,
                               state = 18,
                               county = "163",
                               variables = c(Male_Age_18_19 = "B01001_007",
                                             Male_Age_20 = "B01001_008",
                                             Male_Age_21 = "B01001_009",
                                             Male_Age_22_24 = "B01001_010",
                                             Male_Age_25_29 = "B01001_011",
                                             Male_Age_30_34 = "B01001_012",
                                             Male_Age_35_39 = "B01001_013",
                                             Male_Age_40_44 = "B01001_014",
                                             Male_Age_45_49 = "B01001_015",
                                             Male_Age_50_54 = "B01001_016",
                                             Male_Age_55_59 = "B01001_017",
                                             Male_Age_60_61 = "B01001_018",
                                             Male_Age_62_64 = "B01001_019",
                                             Male_Age_65_66 = "B01001_020",
                                             Male_Age_67_69 = "B01001_021",
                                             Male_Age_70_74 = "B01001_022",
                                             Male_Age_75_79 = "B01001_023",
                                             Male_Age_80_84 = "B01001_024",
                                             Male_Age_85_over = "B01001_025",
                                             Female_Age_18_19 = "B01001_031",
                                             Female_Age_20 = "B01001_032",
                                             Female_Age_21 = "B01001_033",
                                             Female_Age_22_24 = "B01001_034",
                                             Female_Age_25_29 = "B01001_035",
                                             Female_Age_30_34 = "B01001_036",
                                             Female_Age_35_39 = "B01001_037",
                                             Female_Age_40_44 = "B01001_038",
                                             Female_Age_45_49 = "B01001_039",
                                             Female_Age_50_54 = "B01001_040",
                                             Female_Age_55_59 = "B01001_041",
                                             Female_Age_60_61 = "B01001_042",
                                             Female_Age_62_64 = "B01001_043",
                                             Female_Age_65_66 = "B01001_044",
                                             Female_Age_67_69 = "B01001_045",
                                             Female_Age_70_74 = "B01001_046",
                                             Female_Age_75_79 = "B01001_047",
                                             Female_Age_80_84 = "B01001_048",
                                             Female_Age_85_over = "B01001_049"
                               )
)
)

#' Young
Young <- Age_data %>%
  filter(
    variable %in% c(
      "Male_Age_18_19",
      "Female_Age_18_19",
      "Male_Age_20",
      "Female_Age_20",
      "Male_Age_21",
      "Female_Age_21",
      "Male_Age_22_24",
      "Female_Age_22_24",
      "Male_Age_25_29",
      "Female_Age_25_29",
      "Male_Age_30_34",
      "Female_Age_30_34")
  ) %>%
  group_by(GEOID) %>%
  summarise(young = sum(estimate))

#' Older
Older <- Age_data %>%
  filter(
    variable %in% c(
      "Male_Age_55_59",
      "Female_Age_55_59",
      "Male_Age_60_61",
      "Female_Age_60_61",
      "Male_Age_62_64",
      "Female_Age_62_64",
      "Male_Age_65_66",
      "Female_Age_65_66",
      "Male_Age_67_69",
      "Female_Age_67_69",
      "Male_Age_70_74",
      "Female_Age_70_74",
      "Male_Age_75_79",
      "Female_Age_75_79",
      "Male_Age_80_84",
      "Female_Age_80_84",
      "Male_Age_85_over",
      "Female_Age_85_over")
  ) %>%
  group_by(GEOID) %>%
  summarise(older = sum(estimate))

#' Join Census and Age Data
Census_data <- Census_data %>% left_join(Young, by = "GEOID") %>% left_join(Older, by = "GEOID")

#' Interactions
Interactions <- data.frame(get_acs(geography = "tract",
                                   geometry = FALSE,
                                   state = 18,
                                   county = "163",
                                   variables = c(White_Below_Poverty = "B17020A_002",
                                                 Black_Male_55_64 = "B01001B_013",
                                                 Black_Male_65_74 = "B01001B_014",
                                                 Black_Male_75_84 = "B01001B_015",
                                                 Black_Male_85_over = "B01001B_016",
                                                 Black_Female_55_64 = "B01001B_028",
                                                 Black_Female_65_74 = "B01001B_029",
                                                 Black_Female_75_84 = "B01001B_030",
                                                 Black_Female_85_over = "B01001B_031",
                                                 Male_Poverty_18_24 = "B17001_010",
                                                 Male_Poverty_25_34 = "B17001_011",
                                                 Female_Poverty_18_24 = "B17001_024",
                                                 Female_Poverty_25_34 = "B17001_025",
                                                 Male_Poverty_55_64 = "B17001_014",
                                                 Male_Poverty_65_74 = "B17001_015",
                                                 Male_Poverty_75_over = "B17001_016",
                                                 Female_Poverty_55_64 = "B17001_028",
                                                 Female_Poverty_65_74 = "B17001_029",
                                                 Female_Poverty_75_over = "B17001_030",
                                                 Hispanic_Male_55_64 = "B01001I_013",
                                                 Hispanic_Male_65_74 = "B01001I_014",
                                                 Hispanic_Male_75_84 = "B01001I_015",
                                                 Hispanic_Male_85_over = "B01001I_016",
                                                 Hispanic_Female_55_64 = "B01001I_028",
                                                 Hispanic_Female_65_74 = "B01001I_029",
                                                 Hispanic_Female_75_84 = "B01001I_030",
                                                 Hispanic_Female_85_over = "B01001I_031",
                                                 Male_Income_75_100 = "B20001_021",
                                                 Male_Income_100_over = "B20001_022")
)
)

#' White Poverty Interaction
White_Poverty <- Interactions %>%
  filter(variable == "White_Below_Poverty") %>%
  select(GEOID, white_poverty = estimate)

#' Black Older Interaction
Black_Older <- Interactions %>%
  filter(variable %in% c("Black_Male_55_64",
                         "Black_Female_55_64",
                         "Black_Male_65_74",
                         "Black_Female_65_74",
                         "Black_Male_75_84",
                         "Black_Female_75_84",
                         "Black_Male_85_over",
                         "Black_Female_85_over")) %>%
  group_by(GEOID) %>%
  summarise(black_older = sum(estimate))

#' Povery Older Interaction
Poverty_Older <- Interactions %>%
  filter(variable %in% c("Male_Poverty_55_64",
                         "Female_Poverty_55_64",
                         "Male_Poverty_65_74",
                         "Female_Poverty_65_74",
                         "Male_Poverty_75_over",
                         "Female_Poverty_75_over")) %>%
  group_by(GEOID) %>%
  summarise(poverty_older = sum(estimate))

#' Hispanic Older Interaction
Hispanic_Older <- Interactions %>%
  filter(variable %in% c("Hispanic_Male_55_64",
                         "Hispanic_Female_55_64",
                         "Hispanic_Male_65_74",
                         "Hispanic_Female_65_74",
                         "Hispanic_Male_75_84",
                         "Hispanic_Female_75_84",
                         "Hispanic_Male_85_over",
                         "Hispanic_Female_85_over")) %>%
  group_by(GEOID) %>%
  summarise(hispanic_older = sum(estimate))

#' Poverty Young Interaction
Poverty_Young <- Interactions %>%
  filter(variable %in% c("Male_Poverty_18_24",
                         "Female_Poverty_18_24",
                         "Male_Poverty_25_34",
                         "Female_Poverty_25_34")) %>%
  group_by(GEOID) %>%
  summarise(poverty_young = sum(estimate))

#' Male Upper Class Interaction
Male_Upper_Class <- Interactions %>%
  filter(variable %in% c("Male_Income_75_100","Male_Income_100_over")) %>%
  group_by(GEOID) %>%
  summarise(male_upper_class = sum(estimate))

#' Census Interaction Join
Census_data <- Census_data %>%
  left_join(White_Poverty, by = "GEOID") %>%
  left_join(Black_Older, by = "GEOID") %>%
  left_join(Poverty_Older, by = "GEOID") %>%
  left_join(Hispanic_Older, by = "GEOID") %>%
  left_join(Poverty_Young, by = "GEOID") %>%
  left_join(Male_Upper_Class, by = "GEOID")

#' Final Data Frame
eville <- evv %>% mutate(GEOID = as.character(GEOID)) %>% left_join(Census_data, by = "GEOID")
Complete Data Frame
}
\keyword{datasets}
