% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{evv}
\alias{evv}
\title{Reading}
\format{An object of class \code{data.frame} with 49 rows and 53 columns.}
\usage{
evv

evv

evv
}
\description{
Reading

Transform

Frame
}
\keyword{datasets}
