% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{evv_map}
\alias{evv_map}
\title{Get Map}
\format{An object of class \code{ggmap} (inherits from \code{raster}) with 1377 rows and 717 columns.}
\usage{
evv_map
}
\description{
Get Map
}
\keyword{datasets}
