% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{High_School}
\alias{High_School}
\title{High School}
\format{An object of class \code{data.frame} with 49 rows and 2 columns.}
\usage{
High_School
}
\description{
High School
}
\keyword{datasets}
