% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Hispanic_Older}
\alias{Hispanic_Older}
\title{Hispanic Older Interaction}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49 rows and 2 columns.}
\usage{
Hispanic_Older
}
\description{
Hispanic Older Interaction
}
\keyword{datasets}
