% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Income_tract}
\alias{Income_tract}
\title{Income Data}
\format{An object of class \code{data.frame} with 784 rows and 5 columns.}
\usage{
Income_tract
}
\description{
Income Data
}
\keyword{datasets}
