% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Interactions}
\alias{Interactions}
\title{Interactions}
\format{An object of class \code{data.frame} with 1421 rows and 5 columns.}
\usage{
Interactions
}
\description{
Interactions
}
\keyword{datasets}
