% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{log_model_per_person_smoke}
\alias{log_model_per_person_smoke}
\title{Basic Model}
\format{An object of class \code{glm} (inherits from \code{lm}) of length 30.}
\usage{
log_model_per_person_smoke
}
\description{
Basic Model
}
\keyword{datasets}
