% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Male_Upper_Class}
\alias{Male_Upper_Class}
\title{Male Upper Class Interaction}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49 rows and 2 columns.}
\usage{
Male_Upper_Class
}
\description{
Male Upper Class Interaction
}
\keyword{datasets}
