% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{No_GED}
\alias{No_GED}
\title{Sorting Education Data}
\format{An object of class \code{data.frame} with 49 rows and 2 columns.}
\usage{
No_GED
}
\description{
Sorting Education Data
}
\keyword{datasets}
