% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Poverty_Young}
\alias{Poverty_Young}
\title{Poverty Young Interaction}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49 rows and 2 columns.}
\usage{
Poverty_Young
}
\description{
Poverty Young Interaction
}
\keyword{datasets}
