% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\name{prediction_map}
\alias{prediction_map}
\title{Find correct ggplot to return}
\usage{
prediction_map(input1, input2)
}
\arguments{
\item{input1}{Selected response from shiny UI}

\item{input2}{Selected plot type from shiny UI}
}
\value{
correct ggplot
}
\description{
Returns ggplot
}
