% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Upper_Class}
\alias{Upper_Class}
\title{Upper Class}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 49 rows and 2 columns.}
\usage{
Upper_Class
}
\description{
Upper Class
}
\keyword{datasets}
