% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{White_Poverty}
\alias{White_Poverty}
\title{White Poverty Interaction}
\format{An object of class \code{data.frame} with 49 rows and 2 columns.}
\usage{
White_Poverty
}
\description{
White Poverty Interaction
}
\keyword{datasets}
