% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{Census_data}
\alias{Census_data}
\title{Full Census}
\format{An object of class \code{data.frame} with 49 rows and 8 columns.}
\usage{
Census_data

Census_data

Census_data
}
\description{
Full Census

Joining Sex

Census Join Income
}
\keyword{datasets}
