% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{eville}
\alias{eville}
\title{Complete Data Frame}
\format{An object of class \code{data.frame} with 49 rows and 56 columns.}
\usage{
eville

eville

eville
}
\description{
Complete Data Frame

Convert Census Data

Prediction Values
}
\keyword{datasets}
