% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Web_Prep2.R
\docType{data}
\name{eville}
\alias{eville}
\title{Final Data Frame}
\format{An object of class \code{data.frame} with 49 rows and 74 columns.}
\usage{
eville

eville

eville
}
\description{
Final Data Frame

Convert Census Data

Predictions
}
\keyword{datasets}
